+++ src/connectedvmware/azext_connectedvmware/vendored_sdks/aio/operations/__init__.py
@@ -7,12 +7,12 @@
 # --------------------------------------------------------------------------
 
 from ._operations import Operations
+from ._virtual_machines_operations import VirtualMachinesOperations
 from ._resource_pools_operations import ResourcePoolsOperations
 from ._clusters_operations import ClustersOperations
 from ._hosts_operations import HostsOperations
 from ._datastores_operations import DatastoresOperations
 from ._vcenters_operations import VCentersOperations
-from ._virtual_machines_operations import VirtualMachinesOperations
 from ._virtual_machine_templates_operations import VirtualMachineTemplatesOperations
 from ._virtual_networks_operations import VirtualNetworksOperations
 from ._inventory_items_operations import InventoryItemsOperations
@@ -22,12 +22,12 @@ from ._guest_agents_operations import GuestAgentsOperations
 
 __all__ = [
     'Operations',
+    'VirtualMachinesOperations',
     'ResourcePoolsOperations',
     'ClustersOperations',
     'HostsOperations',
     'DatastoresOperations',
     'VCentersOperations',
-    'VirtualMachinesOperations',
     'VirtualMachineTemplatesOperations',
     'VirtualNetworksOperations',
     'InventoryItemsOperations',